+++ impl/memory.c
@@ -24,7 +24,7 @@ attr_c_call attr_export
 void* memmove( void* str1, const void* str2, usize n );
 
 attr_internal
-void internal_memory_set_word( void* dst, u8 val, usize size ) {
+void* internal_memory_set_word( void* dst, u8 val, usize size ) {
 
     usize sizeptr = size / sizeof(usize);
     union { u8 bytes[sizeof(usize)]; usize bytesptr; } valptr;
@@ -50,11 +50,13 @@ void internal_memory_set_word( void* dst, u8 val, usize size ) {
     for( usize i = 0; i < rem; ++i ) {
         *(dst_rem + i) = val;
     }
+
+    return dst;
 }
 
 
 attr_internal
-void internal_memory_copy_word(
+void* internal_memory_copy_word(
     void* attr_restrict dst, const void* attr_restrict src, usize size
 ) {
     usize* dstptr = (usize*)dst;
@@ -80,6 +82,8 @@ internal_memory_copy_bytes:
             *dstbyte++ = *srcbyte++;
         }
     }
+
+    return dst;
 }
 
 #define INTERNAL_MEMORY_COPY_PREFER_SIMD_SIZE (gibibytes(1) + mebibytes(512))
@@ -87,12 +91,11 @@ internal_memory_copy_bytes:
 #if defined(CORE_ENABLE_SSE_INSTRUCTIONS)
 
 attr_always_inline inline attr_internal
-void internal_memory_copy_sse(
+void* internal_memory_copy_sse(
     void* attr_restrict dst, const void* attr_restrict src, usize size
 ) {
     if( size < INTERNAL_MEMORY_COPY_PREFER_SIMD_SIZE ) {
-        internal_memory_copy_word( dst, src, size );
-        return;
+        return internal_memory_copy_word( dst, src, size );
     }
 
     // TODO(alicia): stream and stream_load!
@@ -114,34 +117,36 @@ void internal_memory_copy_sse(
     for( usize i = 0; i < rem; ++i ) {
         *dstbyte++ = *srcbyte++;
     }
+
+    return dst;
 }
 
 #endif /* SSE */
 
 attr_core_api
-void memory_copy(
+void* memory_copy(
     void* attr_restrict dst, const void* attr_restrict src, usize size
 ) {
 #if defined(CORE_ENABLE_SSE_INSTRUCTIONS)
-    internal_memory_copy_sse( dst, src, size );
+    return internal_memory_copy_sse( dst, src, size );
 #else
-    internal_memory_copy_word( dst, src, size );
+    return internal_memory_copy_word( dst, src, size );
 #endif
 }
 attr_core_api
-void memory_move(
+void* memory_move(
     void* dst, const void* src, usize size
 ) {
     // TODO(alicia): implementation with sse
-    (void)memmove( dst, src, size );
+    return memmove( dst, src, size );
 }
 attr_core_api
-void memory_set( void* dst, u8 byte, usize size ) {
+void* memory_set( void* dst, u8 byte, usize size ) {
     // TODO(alicia): implementation with sse
-    internal_memory_set_word( dst, byte, size );
+    return internal_memory_set_word( dst, byte, size );
 }
 attr_core_api
-void memory_set_chunks(
+void* memory_set_chunks(
     void* dst, usize chunk_size, const void* chunk, usize count
 ) {
     // TODO(alicia): SIMD?
@@ -150,6 +155,7 @@ void memory_set_chunks(
         memory_copy( dst_bytes, chunk, chunk_size );
         dst_bytes += chunk_size;
     }
+    return dst;
 }
 attr_core_api
 b32 memory_cmp( const void* a, const void* b, usize size ) {
@@ -190,48 +196,6 @@ void memory_free( void* buffer, usize size ) {
     if( buffer ) {
         platform_heap_free( buffer, (usize)size );
     }
-}
-attr_core_api
-void* memory_alloc_aligned( usize size, usize alignment ) {
-    void* ptr = memory_alloc( size + alignment + sizeof(void*) );
-    if( !ptr ) {
-        return NULL;
-    }
-
-    void* aligned = memory_align( (u8*)ptr + sizeof(void*), alignment );
-    ((void**)aligned)[-1] = ptr;
-
-    return aligned;
-}
-attr_core_api
-void* memory_realloc_aligned(
-    void* old_buffer, usize old_size,
-    usize new_size, usize alignment
-) {
-    void* base = ((void**)old_buffer)[-1];
-    usize off  = (usize)old_buffer - (usize)base;
-
-    void* new_buffer = memory_realloc(
-        base, old_size + alignment + sizeof(void*), new_size );
-    if( !new_buffer ) {
-        return NULL;
-    }
-
-    void* aligned = (void*)((u8*)new_buffer + off);
-    ((void**)aligned)[-1] = new_buffer;
-
-    return aligned;
-}
-attr_core_api
-void memory_free_aligned(
-    void* buffer, usize size, usize alignment
-) {
-    if( !buffer ) {
-        return;
-    }
-
-    void* base = ((void**)buffer)[-1];
-    memory_free( base, size + alignment + sizeof(void*) );
 }
 attr_core_api
 usize memory_calculate_aligned_size( usize size, usize alignment ) {